requirements.txt
semantic-kernel[qdrant]
qdrant-client
python-dotenv
langchain
mcp[cli]
anyio